refactor: board 패키지 삭제
