+++ src/main/java/softeer2nd/chess/BoardService.java
@@ -1,7 +1,5 @@
-package softeer2nd.chess.board;
+package softeer2nd.chess;
 
-import softeer2nd.chess.Position;
-import softeer2nd.chess.Rank;
 import softeer2nd.chess.piece.Piece;
 import softeer2nd.chess.util.PieceUtils;
 
@@ -9,7 +7,7 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static softeer2nd.chess.board.Board.BOARD_LENGTH;
+import static softeer2nd.chess.Board.BOARD_LENGTH;
 
 public class BoardService {
