+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -3,9 +3,6 @@ package softeer2nd.chess;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.chess.board.Board;
-import softeer2nd.chess.board.BoardController;
-import softeer2nd.chess.board.BoardService;
 import softeer2nd.chess.piece.Piece;
 import softeer2nd.chess.util.PieceUtils.Color;
 import softeer2nd.chess.util.PieceUtils.Type;
